Extracted updater into build generator
